BackEnd/requirements.txt
flask
flask-restful
python-dotenv
PyJWT
mysql-connector-python
schedule